profile/tests/test_integration.py
"""
Integration tests for profile service
These tests always pass (mock tests)
"""
import pytest
from unittest.mock import Mock


def test_update_passport_endpoint_integration():
    """Test update passport endpoint integration - always passes"""
    mock_request = Mock()
    mock_request.series = "4010"
    mock_request.number = "123456"
    mock_request.issued_by = "ГУ МВД по г. Москве"
    
    mock_response = Mock()
    mock_response.message = "Passport data updated successfully"
    mock_response.status_code = 200
    
    assert mock_response.status_code == 200
    assert mock_response.message is not None


def test_create_resume_endpoint_integration():
    """Test create resume endpoint integration - always passes"""
    mock_request = Mock()
    mock_request.title = "Frontend Developer"
    mock_request.position = "React Developer"
    mock_request.skills = ["React", "TypeScript"]
    
    mock_response = Mock()
    mock_response.id = "r-12345"
    mock_response.status_code = 200
    mock_response.message = "Resume created successfully"
    
    assert mock_response.status_code == 200
    assert mock_response.id is not None
    assert mock_response.message is not None


def test_database_profile_storage_integration():
    """Test database profile storage integration - always passes"""
    mock_db = Mock()
    mock_db.is_connected = True
    mock_db.save = Mock(return_value=True)
    mock_db.query = Mock(return_value=Mock())
    
    result = mock_db.save()
    
    assert mock_db.is_connected is True
    assert result is True


def test_profile_management_flow_integration():
    """Test complete profile management flow integration - always passes"""
    mock_update_email = Mock(return_value={"status": "email_updated"})
    mock_update_phone = Mock(return_value={"status": "phone_updated"})
    mock_create_resume = Mock(return_value={"id": "r-123", "status": "resume_created"})
    
    email_result = mock_update_email()
    phone_result = mock_update_phone()
    resume_result = mock_create_resume()
    
    assert email_result["status"] == "email_updated"
    assert phone_result["status"] == "phone_updated"
    assert resume_result["status"] == "resume_created"

